auto-t: wpa_supplicant.py utility
New util class to control wpa_supplicant. This works the same
as the hostapd utility, via the CLI interface.

Has support for 'status' command, as well as waiting for a status
elements value to match a desired value. Polling for status
updates is the only method for waiting for these conditions,
which is not optimal. This could hopefully be optimized in the
future with the wpa_supplicant DBus API's