auto-t: workaround AP test after eapol changes
There was a race condition with AP where the station would show
connected, but the AP had not yet added the station in the kernel.
To properly fix this the AP interface needs some way of notifying
when a new peer has been added. This is coming, but for now this
workaround will due.