netdev: choose correct address on NEW_KEY/SET_STATION
With the introduction of adhoc, its not as simple as choosing
aa/spa addresses when setting the keys. Since adhoc acts as
both the authenticator and supplicant we must check how the netdev
address relates to the particular handshake object. A simple
helper was added to choose the correct addressed based on netdev
type and handshake state.

The station/ap mode behaves exactly the same as before.