eap: basic implementation of ERP and key cache
ERP (EAP Reauthentication Protocol) requires the session-ID and EMSK from
a previous successful EAP run. Since the EAP state is cleared out after
the connection succeeds we must save these two values for future runs.

For now, this is very basic caching. The keys are not stored on disk, nor
are any key lifetimes tracked/enforced. Because of this ERP will only
work if IWD has not exited between a successful EAP run and the ERP
session. Future changes will include the above functionality.

The API eap_set_key_material needed to be changed to take the session ID,
whos derivation is method specific. Some methods (e.g. EAP-PWD) have
their own derivation, while others are specified in RFC 5247. Some methods
have no session-ID derivation, and therefore are not supported by ERP.

ERP in its current form is here to only support FILS. ERP is likely not
widespread and there is no easy way to determine if an AP supports ERP
without trying it. Attempting ERP with a non-ERP enabled AP will actually
result in longer connection times since ERP must fail and then full EAP
is done afterwards. For this reason ERP will not be enabled unless being
used with FILS. To accomplish this a new boolean parameter was added
to eap_new, which can enable ERP on that EAP state machine. FILS will use
this, and, if in the future ERP is desired on other connections this
boolean could be tied to a config option.