network: allow for Hotspot/OSEN networks without provisioning
OSEN (Online Sign Up) and Hotspot 2.0 support querying ANQP to discover
information about the APs support for EAP protocols and credentials.

For OSEN we also query the anonymous identity to be used with WFA-TLS.

As it stand now this will only work for Hotspot/OSEN networks, but could
could be expanded to support other EAP types which use a
username/password for authentication, also avoiding the need for the
user to create a provisioning file.