wiphy: add wiphy_generate_address_from_ssid
This will generate a new MAC based on an SSID. This allows the same
generated MAC to be used for any given SSID/adapter even between
boots (assuming the permanent address does not change). The hash
is produced with HMAC-SHA256, where the key is the SSID and the data
is the adapters permanent address.

The mac_randomize_bytes option is still used to limit the number
of bytes generated, though they are not 'random' in this case.