netdev: generate MAC per-SSID
The option 'mac_randomize' can now be set to a value 'ssid'. This
will generate a new MAC address when connecting to a network using
wiphy_genearte_address_from_ssid.

To acomplish this, the adapter needs to be brought down, then the
MAC changed, then brought back up. This results in a link down
event which is now ignored if the event was generated based on a
call to connect.